run grammar generation from npm-scripts.js rathen than gulpfile.js
